-created eventIterator.tsx and extracted functionality from page.tsx
